Incorporate settings branch
